requirements.txt
flask==2.3.2
torch==2.0.1
torchvision==0.15.2
scikit-learn==1.3.0
Pillow==9.4.0
open-clip-torch==2.20.0
numpy==1.24.4
